split core settings component into multiple components
