+++ comp/core/settings/registry/component_mock.go
@@ -5,7 +5,7 @@
 
 //go:build test
 
-package settings
+package registry
 
 // Mock implements mock-specific methods.
 type Mock interface {
