+++ comp/core/settings/registry/registryimpl/settings.go
@@ -3,14 +3,17 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2024-present Datadog, Inc.
 
-// Package settingsimpl implements the interface for the settings component
-package settingsimpl
+// Package registryimpl implements the interface for the settings component
+package registryimpl
 
 import (
+	"sync"
+
 	"go.uber.org/fx"
 
 	"github.com/DataDog/datadog-agent/comp/core/log"
-	"github.com/DataDog/datadog-agent/comp/core/settings"
+	"github.com/DataDog/datadog-agent/comp/core/settings/registry"
+	"github.com/DataDog/datadog-agent/pkg/config/model"
 	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
 )
 
@@ -24,29 +27,30 @@ func Module() fxutil.Module {
 type provides struct {
 	fx.Out
 
-	Comp settings.Component
+	Comp registry.Component
 }
 
 type dependencies struct {
 	fx.In
 
 	Log      log.Component
-	Settings []settings.RuntimeSetting `group:"runtime_setting"`
+	Settings []registry.RuntimeSetting `group:"runtime_setting"`
 }
 
 type settingsRegistry struct {
-	registeredSettings map[string]settings.RuntimeSetting
+	m                  sync.Mutex
+	registeredSettings map[string]registry.RuntimeSetting
 }
 
 // RuntimeSettings returns all runtime configurable settings
-func (s settingsRegistry) RuntimeSettings() map[string]settings.RuntimeSetting {
+func (s settingsRegistry) RuntimeSettings() map[string]registry.RuntimeSetting {
 	return s.registeredSettings
 }
 
 // GetRuntimeSetting returns the value of a runtime configurable setting
 func (s *settingsRegistry) GetRuntimeSetting(setting string) (interface{}, error) {
 	if _, ok := s.registeredSettings[setting]; !ok {
-		return nil, &settings.SettingNotFoundError{Name: setting}
+		return nil, &registry.SettingNotFoundError{Name: setting}
 	}
 	value, err := s.registeredSettings[setting].Get()
 	if err != nil {
@@ -55,8 +59,18 @@ func (s *settingsRegistry) GetRuntimeSetting(setting string) (interface{}, error
 	return value, nil
 }
 
+// SetRuntimeSetting changes the value of a runtime configurable setting
+func (s *settingsRegistry) SetRuntimeSetting(setting string, value interface{}, source model.Source) error {
+	s.m.Lock()
+	defer s.m.Unlock()
+	if _, ok := s.registeredSettings[setting]; !ok {
+		return &registry.SettingNotFoundError{Name: setting}
+	}
+	return s.registeredSettings[setting].Set(value, source)
+}
+
 func newSettings(deps dependencies) provides {
-	registeredSettings := map[string]settings.RuntimeSetting{}
+	registeredSettings := map[string]registry.RuntimeSetting{}
 
 	providedSettings := fxutil.GetAndFilterGroup(deps.Settings)
