+++ comp/core/settings/registry/registryimpl/settings_mock.go
